Move dylib version stuff to proc-macro-srv
