+++ Cargo.toml
@@ -120,6 +120,8 @@ hashbrown = { version = "0.14", features = [
 indexmap = "2.1.0"
 itertools = "0.12.0"
 libc = "0.2.150"
+libloading = "0.8.0"
+memmap2 = "0.5.4"
 nohash-hasher = "0.2.0"
 oorandom = "11.1.3"
 object = { version = "0.33.0", default-features = false, features = [
@@ -143,6 +145,7 @@ smallvec = { version = "1.10.0", features = [
   "const_generics",
 ] }
 smol_str = "0.2.1"
+snap = "1.1.0"
 text-size = "1.1.1"
 tracing = "0.1.40"
 tracing-tree = "0.3.0"
@@ -154,6 +157,7 @@ tracing-subscriber = { version = "0.3.18", default-features = false, features = 
 triomphe = { version = "0.1.10", default-features = false, features = ["std"] }
 url = "2.3.1"
 xshell = "0.2.5"
+
 
 
 # We need to freeze the version of the crate, as the raw-api feature is considered unstable
